+++ crates/arpa-node/src/event/dkg_success.rs
@@ -1,19 +1,19 @@
 use crate::subscriber::DebuggableEvent;
+use alloy::primitives::Address;
 use arpa_core::Group;
-use ethers::types::Address;
 use threshold_bls::group::Curve;
 
 use super::{types::Topic, Event};
 
 #[derive(Clone, Debug)]
 pub struct DKGSuccess<C: Curve> {
-    pub chain_id: usize,
+    pub chain_id: u64,
     pub id_address: Address,
     pub group: Group<C>,
 }
 
 impl<C: Curve> DKGSuccess<C> {
-    pub fn new(chain_id: usize, id_address: Address, group: Group<C>) -> Self {
+    pub fn new(chain_id: u64, id_address: Address, group: Group<C>) -> Self {
         DKGSuccess {
             chain_id,
             id_address,
