+++ crates/arpa-node/src/event/types.rs
@@ -2,21 +2,21 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Clone, Copy, Serialize, Deserialize, PartialEq, Hash, Eq)]
 pub enum Topic {
-    NewBlock(usize),
+    NewBlock(u64),
     NewDKGTask,
     RunDKG,
     DKGPhase,
     DKGSuccess,
     DKGPostProcess,
-    NewRandomnessTask(usize),
+    NewRandomnessTask(u64),
     NewGroupRelayTask,
-    NewGroupRelayConfirmationTask(usize),
-    ReadyToHandleRandomnessTask(usize),
+    NewGroupRelayConfirmationTask(u64),
+    ReadyToHandleRandomnessTask(u64),
     ReadyToHandleGroupRelayTask,
-    ReadyToHandleGroupRelayConfirmationTask(usize),
-    ReadyToFulfillRandomnessTask(usize),
+    ReadyToHandleGroupRelayConfirmationTask(u64),
+    ReadyToFulfillRandomnessTask(u64),
     ReadyToFulfillGroupRelayTask,
-    ReadyToFulfillGroupRelayConfirmationTask(usize),
+    ReadyToFulfillGroupRelayConfirmationTask(u64),
     NodeActivation,
     ProviderReconnection,
 }
